Split IOpipeExecution into a no-op and another which actually records statistics.
